Change project folder name and resolve Rebase issues
