+++ README.md
@@ -2,6 +2,6 @@
 
 This repository contains various examples of deep learning such as:
 
-## RESNET Image Classification
+## Salmon Trout
 
-Image Classification using RESNET. Sagemaker Notebook and training script for using fine-tuning or transfer learning to allow a pre trained network to classify ants and bees. This is essentially a template to do transfer learning and image classification with PyTorch.
+Image Classification using RESNET Final Layer Training. Attempts to differentiate between salmon and trout in photos. See [here](salmon_trout/) for more information.
