--- resnet_image_classification/Pipfile
@@ -1,18 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-jupyterlab = "*"
-
-[packages]
-numpy = "*"
-torch = "*"
-torchvision = "*"
-matplotlib = "*"
-nbdime = "*"
-sklearn = "*"
-
-[requires]
-python_version = "3.7"
