+++ salmon_trout/Pipfile
@@ -1,0 +1,22 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+jupyterlab = "*"
+bandit = "*"	
+scikit-learn = "*"
+
+[packages]
+numpy = "*"
+torch = "*"
+torchvision = "*"
+matplotlib = "*"
+nbdime = "*"
+sklearn = "*"
+azureml-sdk = {extras = ["automl", "notebooks"],version = "*"}	
+seaborn = "*"
+
+[requires]
+python_version = "3.7"
